[Deployment Graph] Move files out of `pipeline` folder (#25630)
