+++ python/ray/serve/deployment_graph.py
@@ -44,7 +44,7 @@ class RayServeDAGHandle:
         # NOTE: There's nothing user can do about these warnings, we should hide it.
         with _mute_sync_handle_warnings():
             if self.dag_node is None:
-                from ray.serve.pipeline.json_serde import dagnode_from_json
+                from ray.serve.json_serde import dagnode_from_json
 
                 self.dag_node = json.loads(
                     self.dag_node_json, object_hook=dagnode_from_json
