--- python/ray/serve/pipeline/constants.py
@@ -1,5 +1,0 @@
-# Reserved constant used as key in other_args_to_resolve to configure if we
-# return sync or async handle of a deployment.
-# True -> RayServeSyncHandle
-# False -> RayServeHandle
-USE_SYNC_HANDLE_KEY = "__use_sync_handle__"
