+++ python/ray/serve/tests/test_deployment_node.py
@@ -2,7 +2,7 @@ import pytest
 
 import ray
 from ray import serve
-from ray.serve.pipeline.deployment_node import (
+from ray.serve.deployment_node import (
     DeploymentNode,
 )
