+++ python/ray/serve/tests/test_pipeline_dag_plot.py
@@ -12,7 +12,7 @@ from ray.serve.deployment_graph import InputNode
 from ray.serve.drivers import DAGDriver
 import starlette.requests
 
-from ray.serve.pipeline.generate import transform_ray_dag_to_serve_dag
+from ray.serve.generate import transform_ray_dag_to_serve_dag
 
 NESTED_HANDLE_KEY = "nested_handle"
