+++ python/ray/serve/tests/test_pipeline_handle_serde.py
@@ -2,7 +2,7 @@ import sys
 
 import pytest
 from ray.serve.dag import InputNode
-from ray.serve.pipeline.api import build as pipeline_build
+from ray.serve.deployment_graph_build import build as pipeline_build
 
 import ray
 from ray import serve
