+++ python/ray/serve/tests/test_pipeline_ingress_deployment.py
@@ -1,8 +1,8 @@
 import pytest
 import sys
 
-from ray.serve.pipeline.api import get_and_validate_ingress_deployment
-from ray.serve.pipeline.api import build as pipeline_build
+from ray.serve.deployment_graph_build import get_and_validate_ingress_deployment
+from ray.serve.deployment_graph_build import build as pipeline_build
 from ray.serve.dag import InputNode
 from ray import serve
